+++ internal/kafeman/consume.go
@@ -42,6 +42,7 @@ func (k *kafeman) Consume(ctx context.Context, cmd ConsumeCommand, decoder Decod
 
 	messages, err := c.StartConsume(ctx)
 	if err != nil {
+		// TODO: add possible errors
 		return nil, ErrNoTopicProvided
 	}
